refactor(mcp): make every module a leaf or a handler, with no cycles
`pnpm knip` read the split and found what a green test run could not: three
import cycles among the tools modules, and 78 exports nothing imports. A cycle
means the cut was drawn through a dependency rather than around it, and an
export nothing reads is a module pretending to have an interface.

Four helpers move to where both sides can reach them without a back edge:
`assertScanRootAllowed` to `server/runtime.mjs`, which already owns the roots it
guards; `buildSummaryFreshness`, `listVaultSourcePaths` and the whole-vault
issue finders to `tools/vault-nodes.mjs`, which already answers whether a
reference resolves. The relation-key resolution leaves the write module for
`tools/relation-keys.mjs`, so a read handler no longer imports from a writer to
learn which frontmatter key holds an edge. Every import is then regenerated from
the export map rather than patched, and each export list trimmed to what is
actually imported. Result: 0 cycles, knip clean.

Proof: initialize instructions and tools/list still byte-identical to the
pre-split capture, docs:surface:check current at 38 tools, test:mcp:unit
976/976, test:contracts 2646/2646, pnpm lint and pnpm knip clean.
