+++ mcp/src/server/instance.mjs
@@ -3,21 +3,20 @@
  * the live tool inventory. It lives apart from the entry point so a handler can
  * ask the connected client who it is without importing `index.js`.
  */
-import { SERVER_VERSION } from '../server-version.mjs';
-import { Server } from '@modelcontextprotocol/server';
 
+import { SERVER_VERSION } from '../server-version.mjs';
 import { buildToolInventorySection } from '../tool-inventory.mjs';
 import {
   SERVER_INSTRUCTIONS_TEMPLATE,
   TOOL_INVENTORY_PLACEHOLDER,
 } from './instructions.mjs';
 import { TOOLS_FOR_LIST } from './registry.mjs';
+import { Server } from '@modelcontextprotocol/server';
 
 const SERVER_INSTRUCTIONS = SERVER_INSTRUCTIONS_TEMPLATE.replace(
   TOOL_INVENTORY_PLACEHOLDER,
   buildToolInventorySection(TOOLS_FOR_LIST),
 );
-
 
 const server = new Server(
   { name: 'ontology-atlas-mcp', version: SERVER_VERSION },
@@ -28,6 +27,5 @@ const server = new Server(
 );
 
 export {
-  SERVER_INSTRUCTIONS,
   server,
 };
