+++ mcp/src/server/tool-schemas.mjs
@@ -5,6 +5,7 @@
  * Pure data: nothing here reads the disk or the environment, so the public
  * surface a client negotiates is decided by this file and `registry.mjs` alone.
  */
+
 import {
   CONSTRUCTION_ADMISSION_CONTRACT,
   CONSTRUCTION_ADMISSION_TIERS,
@@ -2108,7 +2109,6 @@ const GIT_RESULT_OUTPUT_SCHEMA = Object.freeze({
   additionalProperties: false,
 });
 
-
 const GIT_SNAPSHOT_OUTPUT_SCHEMA = Object.freeze({
   ...GIT_RESULT_OUTPUT_SCHEMA,
   properties: {
@@ -2160,7 +2160,6 @@ const GIT_HISTORY_OUTPUT_SCHEMA = Object.freeze({
 
 export {
   NON_BLANK_STRING_SCHEMA,
-  BACKLINK_REWRITE_VALUE_OUTPUT_SCHEMA,
   GRAPH_REF_ARRAY_MAX_ITEMS,
   LOCALE_LABELS_SCHEMA,
   IGNORE_ARRAY_MAX_ITEMS,
@@ -2171,26 +2170,14 @@ export {
   REVIEW_REQUIRED_CAPABILITY_ROW_SCHEMA,
   PROPOSED_BUSINESS_CONCEPT_ROW_SCHEMA,
   SEMANTIC_EVIDENCE_ROW_SCHEMA,
-  RUST_FEATURE_REFERENCE_OUTPUT_SCHEMA,
   RUST_FEATURE_CONFIGURATION_EVIDENCE_OUTPUT_SCHEMA,
   IMPORT_SCAN_COVERAGE_OUTPUT_SCHEMA,
-  GO_PACKAGE_IMPORT_ROW_SCHEMA,
-  GO_PACKAGE_IMPORT_MODULE_EDGE_SCHEMA,
   GO_PACKAGE_IMPORT_EVIDENCE_OUTPUT_SCHEMA,
   GO_PACKAGE_IMPORT_EVIDENCE_SUMMARY_SCHEMA,
-  MEANING_PROPOSAL_CONCEPT_INPUT_PROPERTIES,
-  COMPETENCY_RELATION_WITNESS_SCHEMA,
-  COMPETENCY_WITNESSES_SCHEMA,
-  COMPETENCY_ANSWER_SCHEMA,
-  COMPETENCY_ANSWERS_SCHEMA,
   MEANING_PROPOSAL_INPUT_SCHEMA,
-  MEANING_WRITE_PLAN_OUTPUT_SCHEMA,
-  CONSTRUCTION_LIFECYCLE_OUTPUT_SCHEMA,
   MEANING_PROPOSAL_VALIDATION_OUTPUT_SCHEMA,
   EXTRACTION_CONTRACT_OUTPUT_SCHEMA,
   RELATION_ARRAY_PATCH_SCHEMA,
-  BACKLINK_REWRITE_KEY_CHANGE_OUTPUT_SCHEMA,
-  BACKLINK_REWRITE_UPDATE_OUTPUT_SCHEMA,
   BACKLINK_REWRITE_PLAN_OUTPUT_SCHEMA,
   BACKLINK_ROW_OUTPUT_SCHEMA,
   CAPTURED_DOC_OUTPUT_SCHEMA,
@@ -2201,14 +2188,9 @@ export {
   BODY_DELIVERY_MODES,
   BODY_INFO_OUTPUT_SCHEMA,
   GROWTH_HINT_OUTPUT_SCHEMA,
-  PROJECT_SOURCE_GAP_SCHEMA,
-  PROJECT_SOURCE_ACTION_SCHEMA,
   PROJECT_SOURCE_RECEIPT_SCHEMA,
   PROJECT_SOURCE_BINDING_VIEW_SCHEMA,
   PROJECT_SOURCE_VIEW_SCHEMA,
-  PROJECT_SOURCE_TOOL_CALL_SCHEMA,
-  PROJECT_SOURCE_CLI_CALL_SCHEMA,
-  PROJECT_SOURCE_UNDO_SCHEMA,
   PROJECT_SOURCE_REMEDY_SCHEMA,
   PROJECT_SOURCE_NEXT_CALL_SCHEMA,
   RELATION_RESULT_SCHEMA,
@@ -2220,12 +2202,6 @@ export {
   NODE_KIND_DESCRIPTION,
   EDGE_TARGET_KIND_DESCRIPTION,
   POST_WRITE_MAINTENANCE_GUIDANCE,
-  COMPACT_MAINTENANCE_PROPOSED_ACTION_TOOLS,
-  COMPACT_MAINTENANCE_PROPOSED_ACTION_ARGS_OUTPUT_SCHEMA,
-  COMPACT_MAINTENANCE_NODE_OUTPUT_SCHEMA,
-  COMPACT_MAINTENANCE_PROPOSED_ACTION_OUTPUT_SCHEMA,
-  COMPACT_MAINTENANCE_ACTION_OUTPUT_SCHEMA,
-  NULLABLE_COMPACT_MAINTENANCE_ACTION_OUTPUT_SCHEMA,
   POST_WRITE_MAINTENANCE_OUTPUT_SCHEMA,
   MEANING_ASSESSMENT_OUTPUT_SCHEMA,
   nonBlankStringSchema,
@@ -2234,9 +2210,6 @@ export {
   QUERY_PLAN_TARGET_OPERATION_UNION,
   RELATION_TYPE_UNION,
   ADD_RELATION_TYPE_SCHEMA,
-  GIT_FILE_OUTPUT_SCHEMA,
-  GIT_COUNTS_OUTPUT_SCHEMA,
-  GIT_RISK_OUTPUT_SCHEMA,
   DESTRUCTIVE_PREVIEW_OUTPUT_PROPERTIES,
   DESTRUCTIVE_PREVIEW_REQUIRED,
   GIT_RESULT_OUTPUT_SCHEMA,
