+++ mcp/src/tools/repo-analysis.mjs
@@ -3,7 +3,8 @@
  * `inspect_architecture`, `infer_imports`, `index_project` — plus the scan-root
  * guard that keeps them inside the vault's own repository.
  */
-import { analyzeRepoStructure } from '../analyze.mjs';
+
+import { analyzeRepoStructure } from '../analyze/repo-structure.mjs';
 import {
   buildArchitectureBrief,
   buildArchitectureMeasuredStamp,
@@ -24,6 +25,7 @@ import {
   COMPILED_ONTOLOGY_CACHE,
   REPO_ROOT,
   VAULT_ROOT,
+  assertScanRootAllowed,
 } from '../server/runtime.mjs';
 import {
   IGNORE_ARRAY_MAX_ITEMS,
@@ -42,66 +44,9 @@ import {
 import { loadVaultDocs } from '../vault.mjs';
 import { validateVaultTool } from './validate-vault.mjs';
 import {
-  existsSync,
-  realpathSync,
-} from 'node:fs';
-import {
-  isAbsolute,
   relative,
-  resolve,
   sep,
 } from 'node:path';
-
-// Thin wrapper over analyze_repo_structure. Zero side effects — it never touches
-// vault frontmatter. Only the exact writePlan returned after reviewPlan plus
-// independent qualification is a truth entry point for the batch writer.
-/**
- * Is this a place we may scan — **it must be inside the vault or its repository.**
- *
- * **Why** (review 2026-08-16, confirmed by measurement): `analyze_repo_structure`,
- * `infer_imports`, `index_project`, and `validate_vault` took a `rootPath` (or
- * `repoRoot`), called `resolve()` on it, and **checked no boundary at all**. So
- * this call succeeded as written:
- *
- * ```
- * analyze_repo_structure {"rootPath":"/etc"}  → ok, returns the directory structure
- * ```
- *
- * Worse, all four are **read tools**, so `OATLAS_READ_ONLY` does not stop them.
- * That mode is recommended when whoever registered the server is not the vault's
- * owner — and it left them unable to write but **able to scan the entire disk**.
- *
- * This collides head-on with what the product promises its users: *"files on the
- * user's disk such as passwords or credentials are never scanned automatically"*
- * (`.claude/rules/local-first.md`), *"we do not scan the user's disk
- * automatically"* (the trust charter). A tool call steered by one line of prompt
- * would break that promise.
- *
- * So only the vault, or that vault's repository, is allowed. Real paths are
- * resolved before comparison to close the symlink escape — the same grammar
- * `absorb_document` already uses.
- */
-function assertScanRootAllowed(target, argName = 'rootPath') {
-  const canonical = existsSync(target) ? realpathSync(target) : resolve(target);
-  const roots = [];
-  for (const root of [VAULT_ROOT, REPO_ROOT]) {
-    try {
-      roots.push(existsSync(root) ? realpathSync(root) : resolve(root));
-    } catch {
-      roots.push(resolve(root));
-    }
-  }
-  const inside = roots.some((root) => {
-    if (canonical === root) return true;
-    const rel = relative(root, canonical);
-    return rel !== '' && rel !== '..' && !rel.startsWith(`..${sep}`) && !isAbsolute(rel);
-  });
-  if (inside) return canonical;
-  throw new Error(
-    `${argName} must be inside the vault (${roots[0]}) or its repository (${roots[1]}). ` +
-      'This server only reads the folder it was opened for.',
-  );
-}
 
 function analyzeRepoStructureTool({ rootPath, maxDepth, ignore, proposal, qualification } = {}) {
   requireOptionalNonBlankString(rootPath, 'rootPath');
@@ -777,15 +722,8 @@ function summarizeProposedBusinessConceptRows(rows) {
 }
 
 export {
-  assertScanRootAllowed,
   analyzeRepoStructureTool,
   inspectArchitectureTool,
-  buildImportStaleEdgeFollowUp,
-  buildGoPackageImportEvidenceSummary,
   inferImportsTool,
-  estimateMcpToolResultUtf8Bytes,
   indexProjectTool,
-  summarizeBusinessEvidenceRows,
-  summarizeReviewRequiredCapabilityRows,
-  summarizeProposedBusinessConceptRows,
 };
